Improved GraphQl server
