--- src/graphql/resolvers/contact-resolver.js
@@ -1,15 +1,0 @@
-import { buildSchema } from 'graphql';
-
-export const schema = buildSchema(`
-	type Query {
-    contact(id: String!): Contact
-    contacts: [Contact]
-  },
-
-  type Contact {
-    id: String
-    name: String
-    email: String
-    phone: String
-  }
-`);
