--- src/graphql/resolvers/index.js
@@ -1,5 +1,0 @@
-import contactResolver from './contact-resolver';
-
-export default {
-	contactResolver
-};
